+++ src/stuart/util/isInstanceOf.lua
@@ -1,3 +1,3 @@
-local isInstanceOf = function(obj, class)
-  return obj.isInstanceOf ~= nil and obj:isInstanceOf(class) end
-return isInstanceOf
+return function(obj, class)
+  return type(obj) == 'table' and obj.isInstanceOf and obj:isInstanceOf(class)
+end